# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Garage Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  loginCode: string;
  role: UserRole;
  /** @nullable */
  deviceId?: string | null;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  username: string;
  /** @minLength 4 */
  loginCode: string;
  deviceId?: string;
}

export interface LoginInput {
  loginCode: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface CustomerRecord {
  id: number;
  userId: number;
  username: string;
  firstName: string;
  lastName: string;
  breakdownType: string;
  totalAmount: number;
  /** @nullable */
  customerNumber?: string | null;
  carType: string;
  licensePlate: string;
  createdAt: string;
}

export interface RecordInput {
  firstName: string;
  lastName: string;
  breakdownType: string;
  totalAmount: number;
  customerNumber?: string;
  carType: string;
  licensePlate: string;
}

export interface RecordUpdate {
  firstName?: string;
  lastName?: string;
  breakdownType?: string;
  totalAmount?: number;
  customerNumber?: string;
  carType?: string;
  licensePlate?: string;
}

export interface Invoice {
  id: number;
  userId: number;
  username: string;
  /** @nullable */
  customerNumber?: string | null;
  firstName: string;
  lastName: string;
  carType: string;
  licensePlate: string;
  workshopName: string;
  breakdownType: string;
  paymentMethod: string;
  amount: number;
  createdAt: string;
}

export interface InvoiceInput {
  customerNumber?: string;
  firstName: string;
  lastName: string;
  carType: string;
  licensePlate: string;
  workshopName: string;
  breakdownType: string;
  paymentMethod: string;
  amount: number;
}

export interface InvoiceUpdate {
  customerNumber?: string;
  firstName?: string;
  lastName?: string;
  carType?: string;
  licensePlate?: string;
  workshopName?: string;
  breakdownType?: string;
  paymentMethod?: string;
  amount?: number;
}

export interface Stats {
  totalRecords: number;
  totalInvoices: number;
  totalAmount: number;
  recentRecords: CustomerRecord[];
}

export type ListRecordsParams = {
search?: string;
deviceId?: string;
};

export type ListInvoicesParams = {
search?: string;
deviceId?: string;
};