# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Garage Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Invoice {
  id: number;
  userId: number;
  username: string;
  /** @nullable */
  customerNumber?: string | null;
  firstName: string;
  lastName: string;
  carType: string;
  licensePlate: string;
  workshopName: string;
  breakdownType: string;
  paymentMethod: string;
  amount: number;
  createdAt: string;
}