# lib/api-zod/src/generated/types/invoiceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Garage Management API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceUpdate {
  customerNumber?: string;
  firstName?: string;
  lastName?: string;
  carType?: string;
  licensePlate?: string;
  workshopName?: string;
  breakdownType?: string;
  paymentMethod?: string;
  amount?: number;
}