# lib/api-zod/src/generated/types/recordInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Garage Management API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordInput {
  firstName: string;
  lastName: string;
  breakdownType: string;
  totalAmount: number;
  customerNumber?: string;
  carType: string;
  licensePlate: string;
}