# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Garage Management API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 2 */
  username: string;
  /** @minLength 4 */
  loginCode: string;
  deviceId?: string;
}